refactor(Modpath): refactor Modpath to Modpath6
Refactor to improve use of glob for creating documentation. Also
reorganized tutorials so directory structure can be used to develop
tutorial rst file.
