+++ flopy/modpath/__init__.py
@@ -1,6 +1,6 @@
-from .mp import Modpath
-from .mpbas import ModpathBas
-from .mpsim import ModpathSim
+from .mp6 import Modpath6
+from .mp6bas import Modpath6Bas
+from .mp6sim import Modpath6Sim
 from .mp7 import Modpath7
 from .mp7bas import Modpath7Bas
 from .mp7sim import Modpath7Sim
